+++ spec/send_headers.spec.php
@@ -11,6 +11,7 @@ describe(\CacheControl\SendHeaders::class, function () {
 		allow('is_front_page')->toBeCalled()->andReturn(false);
 		allow('is_home')->toBeCalled()->andReturn(false);
 		allow('is_user_logged_in')->toBeCalled()->andReturn(false);
+		allow('is_preview')->toBeCalled()->andReturn(false);
 		allow('get_post_type')->toBeCalled()->andReturn('post');
 		allow('get_post_taxonomies')->toBeCalled()->andReturn(['category', 'post_tag', 'custom-taxonomy']);
 		allow('get_page_template_slug')->toBeCalled()->andReturn('default');
@@ -148,6 +149,16 @@ describe(\CacheControl\SendHeaders::class, function () {
 				allow('is_user_logged_in')->toBeCalled()->andReturn(true);
 			});
 
+			it('sets a private cache header for preview pages', function () {
+				allow('is_preview')->toBeCalled()->andReturn(true);
+				expect('get_field')->toBeCalled()->once();
+
+				allow('header')->toBeCalled();
+				expect('header')->toBeCalled()->once()->with('Cache-Control: no-cache, no-store, private');
+
+				$this->sendHeaders->setCacheHeader();
+			});
+
 			it('is not in developer mode', function () {
 				expect('get_field')->toBeCalled()->once();
 
@@ -178,6 +189,18 @@ describe(\CacheControl\SendHeaders::class, function () {
 			});
 		});
 
+		context('the user is not logged in', function () {
+			it('sets a private cache header for preview pages', function () {
+				allow('is_preview')->toBeCalled()->andReturn(true);
+				expect('get_field')->toBeCalled()->once();
+
+				allow('header')->toBeCalled();
+				expect('header')->toBeCalled()->once()->with('Cache-Control: no-cache, no-store, private');
+
+				$this->sendHeaders->setCacheHeader();
+			});
+		});
+
 		context('Testing developer-mode', function () {
 			beforeEach(function () {
 				allow('is_user_logged_in')->toBeCalled()->andReturn(true);
@@ -187,7 +210,7 @@ describe(\CacheControl\SendHeaders::class, function () {
 
 			it('is in developer mode on the local dev environment', function () {
 				expect('get_post_types')->toBeCalled()->once();
-				expect('get_field')->toBeCalled()->times(3);
+				expect('get_field')->toBeCalled()->times(1);
 
 				allow('header')->toBeCalled();
 				expect('header')->toBeCalled()->once()->with('Meta-cc-post-type: post');
@@ -200,12 +223,18 @@ describe(\CacheControl\SendHeaders::class, function () {
 				expect('header')->toBeCalled()->once()->with('Meta-cc-template_name: default');
 				expect('header')->toBeCalled()->once()->with('Meta-cc-requires-password: no');
 				expect('header')->toBeCalled()->once()->with('Meta-cc-post-types: post,page,custom-post');
-				expect('header')->toBeCalled()->once()->with('Meta-cc-front-page-cache-value: default');
-				expect('header')->toBeCalled()->once()->with('Meta-cc-configured-max-age: 86400');
-				expect('header')->toBeCalled()->once()->with('Meta-cc-configured-cache: no-cache (logged in user)');
-				expect('header')->toBeCalled()->once()->with('Meta-cc-currently-used-config: default');
-				expect('header')->toBeCalled()->once()->with('Meta-cc-final-configured-max-age: 86400');
-				expect('header')->toBeCalled()->once()->with('Cache-Control: max-age=86400, public');
+				expect('header')->toBeCalled()->once()->with('Cache-Control: no-cache, no-store, private');
+
+				$this->sendHeaders->setCacheHeader();
+			});
+
+			it('is on the local environment', function () {
+				allow('wp_get_environment_type')->toBeCalled()->andReturn('local');
+				expect('get_post_types')->toBeCalled()->once();
+				expect('get_field')->toBeCalled()->times(1);
+
+				allow('header')->toBeCalled();
+				expect('header')->toBeCalled()->times(12);
 
 				$this->sendHeaders->setCacheHeader();
 			});
@@ -213,21 +242,18 @@ describe(\CacheControl\SendHeaders::class, function () {
 			it('is on the development environment', function () {
 				allow('wp_get_environment_type')->toBeCalled()->andReturn('development');
 				expect('get_post_types')->toBeCalled()->once();
-				expect('get_field')->toBeCalled()->times(3);
+				expect('get_field')->toBeCalled()->times(1);
 
 				allow('header')->toBeCalled();
-				expect('header')->toBeCalled()->times(17);
+				expect('header')->toBeCalled()->times(12);
 
 				$this->sendHeaders->setCacheHeader();
 			});
 
 			it('is on the staging environment', function () {
 				allow('wp_get_environment_type')->toBeCalled()->andReturn('staging');
-				expect('get_post_types')->toBeCalled()->once();
-				expect('get_field')->toBeCalled()->times(3);
-
 				allow('header')->toBeCalled();
-				expect('header')->toBeCalled()->times(17);
+				expect('header')->toBeCalled()->once()->with('Cache-Control: no-cache, no-store, private');
 
 				$this->sendHeaders->setCacheHeader();
 			});
@@ -254,10 +280,10 @@ describe(\CacheControl\SendHeaders::class, function () {
 				allow('get_page_template_slug')->toBeCalled()->andReturn('custom-template.php');
 
 				expect('get_post_types')->toBeCalled()->once();
-				expect('get_field')->toBeCalled()->times(3);
+				expect('get_field')->toBeCalled()->times(1);
 
 				allow('header')->toBeCalled();
-				expect('header')->toBeCalled()->times(27);
+				expect('header')->toBeCalled()->times(12);
 
 				$this->sendHeaders->setCacheHeader();
 			});